requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
astroid==2.9.0
async-timeout==4.0.2
asyncio==3.4.3
asyncpg==0.25.0
attrs==21.4.0
black==21.11b1
boto3==1.21.10
botocore==1.24.10
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
frozenlist==1.3.0
greenlet==1.1.2
httptools==0.3.0
idna==3.3
isort==5.10.1
jmespath==0.10.0
lazy-object-proxy==1.6.0
marshmallow==3.14.1
mccabe==0.6.1
multidict==5.2.0
mypy==0.910
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
Pygments==2.10.0
pylint==2.12.2
python-dateutil==2.8.2
python-dotenv==0.19.2
PyYAML==6.0
regex==2021.11.10
rich==10.15.2
s3transfer==0.5.2
sanic==21.9.3
sanic-ext==22.1.2
sanic-routing==0.7.2
Shapely==1.8.1.post1
six==1.16.0
SQLAlchemy==1.4.27
sqlalchemy2-stubs==0.0.2a19
toml==0.10.2
tomli==1.2.2
typing_extensions==4.0.1
ujson==5.1.0
urllib3==1.26.8
uvloop==0.16.0
websockets==10.1
wrapt==1.13.3
yarl==1.7.2
